+++ .gitmodules
@@ -5,3 +5,6 @@
 [submodule "submodules/finance_agent"]
 	path = submodules/finance_agent
 	url = git@github.com:vals-ai/finance-agent.git
+[submodule "services/benchmarks/swebench"]
+	path = services/benchmarks/swebench
+	url = git@github.com:vals-ai/SWEBenchBenchmarkService.git
