+++ cli/src/cli/__init__.py
@@ -1,0 +1,2 @@
+def hello() -> str:
+    return "Hello from cli!"
