+++ datasets/fab/dataset.py
@@ -4,9 +4,9 @@ from model_library.base import TextInput
 from typing_extensions import override
 from vals import Suite
 
-from src.base.dataset import Dataset
-from src.base.types import Task, TaskGroup
-from src.logger import get_logger
+from agentic_harness.base.dataset import Dataset
+from agentic_harness.base.types import Task, TaskGroup
+from agentic_harness.logger import get_logger
 
 logger = get_logger(__name__)
