+++ src/agentic_harness/base/__init__.py
