+++ src/agentic_harness/base/dataset.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from src.base.types import DatasetConfig, TaskGroup
+from agentic_harness.base.types import DatasetConfig, TaskGroup
 
 
 class Dataset(ABC):
