+++ src/agentic_harness/base_agent.py
@@ -1,7 +1,8 @@
 from model_library.base import QueryResult
-from src.base.evaluate import Evaluator
-from src.base.types import AgentConfig, Task
-from src.base.contract import AgentContract
+
+from agentic_harness.base.contract import AgentContract
+from agentic_harness.base.evaluate import Evaluator
+from agentic_harness.base.types import AgentConfig, Task
 
 
 class BaseAgent:
