+++ src/agentic_harness/registry.py
@@ -2,17 +2,17 @@ from importlib import import_module
 from inspect import isclass
 from typing import TypeVar
 
-from src.base.benchmark import Benchmark
-from src.base.contract import AgentContract
-from src.base.dataset import Dataset
-from src.base.types import BaseConfig
-from src.base_agent import BaseAgent
-from src.evaluators.platform_evaluate import PlatformEvaluator
-from src.logger import get_logger
+from agentic_harness.base.benchmark import Benchmark
+from agentic_harness.base.contract import AgentContract
+from agentic_harness.base.dataset import Dataset
+from agentic_harness.base.types import BaseConfig
+from agentic_harness.base_agent import BaseAgent
+from agentic_harness.evaluators.platform_evaluate import PlatformEvaluator
+from agentic_harness.logger import get_logger
 
 logger = get_logger(__name__)
 
-BENCHMARK_PACKAGE = "src.benchmarks"
+BENCHMARK_PACKAGE = "agentic_harness.benchmarks"
 BENCHMARK_MODULE = "benchmark"
 DATASET_PACKAGE = "datasets"
 DATASET_MODULE = "dataset"
